+++ docs/changelog.rst
@@ -8,6 +8,8 @@ ChangeLog
 
 - made urls Django 1.5 compatible
 - add maximum_choices field
+- drop unique constraint on field label
+
 0.4.2
 -----